refactor: add useGamePlay custom hook
게임플레이 페이지에서 사용하는 상태변수들을 한번에 선언해서 내보내주는 게임플레이페이지용 커스텀훅 생성
